Add cmd directory to go services
