+++ services/public-api/Dockerfile
@@ -14,5 +14,5 @@ COPY migrations ./migrations
 
 COPY . .
 
-RUN go build -mod=vendor -o main ./api/api.go
+RUN go build -mod=vendor -o main ./cmd/main.go
 CMD ["./main", "public-api", "api-store"]
