+++ services/public-api/cmd/main.go
@@ -1,0 +1,18 @@
+package main
+
+import (
+	"os"
+	"web.microservice.shell/services/public-api/api"
+)
+
+func main() {
+	host := os.Args[1]
+	if host == "" {
+		panic("Host not defined")
+	}
+	store := os.Args[2]
+	if store == "" {
+		panic("Store not defined")
+	}
+	api.Run(store)
+}
